Fix invalid CSRF token when there are multiple cookies (#267)
The logic for pulling the CSRF token out from the cookies did
not account for multiple cookies, and whitespace between them, e.g.

	cookieA=value1; cookieB=value2

Consequently, when the CSRF token was pulled from the cookies,
the slice would not have accounted for the extra space,
e.g. ` csrf-token=abcdef` => `=abcdef` as the final CSRF token.